requirements.txt
tkinter
pillow
matplotlib
opencv-python
random
numpy